Move UploadSentinel from postgres to internal
